refactor: rename profile_screen_widgets.dart to other_profile_screen_widgets.dart
