+++ lib/view/profile/profile_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:simple_notes_app/view/authentication/authentication.dart';
-import 'package:simple_notes_app/view/profile/profile_screen_widgets.dart';
+import 'package:simple_notes_app/view/profile/widgets/other_profile_screen_widgets.dart';
 import 'package:simple_notes_app/view/profile/widgets/profile_card.dart';
 import 'package:simple_notes_app/widgets_library/widgets_library.dart';
